Rework to (hopefully) be more readable
